+++ src/resources/views/index.blade.php
@@ -12,7 +12,7 @@
                     <a href="">選考対策</a>
                     <a href="{{ route('research.index') }}">業界研究</a>
                     <a href="{{ route('reserve.index') }}">予約</a>
-                    <a href="{{ route('profile.index') }}">マイページ</a>
+                    <a href="{{ route('mypage.index') }}">マイページ</a>
                 </div>
             </div>
         </div>
