+++ src/resources/views/layouts/navigation.blade.php
@@ -34,7 +34,7 @@
                     </x-slot>
 
                     <x-slot name="content">
-                        <x-dropdown-link :href="route('profile.index')">
+                        <x-dropdown-link :href="route('mypage.index')">
                             {{ __('Profile') }}
                         </x-dropdown-link>
 
@@ -80,7 +80,7 @@
             </div>
 
             <div class="mt-3 space-y-1">
-                <x-responsive-nav-link :href="route('profile.index')">
+                <x-responsive-nav-link :href="route('mypage.index')">
                     {{ __('Profile') }}
                 </x-responsive-nav-link>
